requirements.txt
tensorflow>=2.12
numpy
opencv-python
pillow
matplotlib
scikit-learn
